Rename LinkComponent to ProfileLink to match its inner exported function
