+++ client/src/components/ProfileComps/RewardsPage.js
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import NavigationBar from "../NavBar/NavigationBar";
 import { useAuth } from "../../database/auth";
-import ProfileLink from "./LinkComponent";
+import ProfileLink from "./ProfileLink";
 import AvatarCard from "./AvatarCard";
 
 const MainContainer = styled.div`
